shop/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.0
cron-descriptor==1.4.0
Django==4.2.3
django-celery-beat==2.5.0
django-timezone-field==6.0
djangorestframework==3.14.0
hiredis==2.2.3
idna==3.4
kavenegar==1.1.2
kombu==5.3.1
linecache2==1.0.0
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
traceback2==1.4.0
tzdata==2023.3
unittest2==1.1.0
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
